refactor: แก้ปัญหาเรื่อง Token และลบ lib ที่ไม่ได้ใช้แล้ว

+++ lib/Home/Knowledge/PageKnowledge.dart
@@ -187,7 +187,7 @@ class _KnowledgePageState extends State<KnowledgePage> {
                           dotsCount: widget.knowledge != null
                               ? widget.knowledge!.knowledgeImg.length
                               : widget.contents!.ImageURL.length,
-                          position: currentIndex,
+                          position: currentIndex.toDouble(),
                           onTap: (position) {
                             controller.animateToPage(position.toInt());
                           },
